+++ models/flavor.py
@@ -8,7 +8,7 @@ from models.model_base import ModelBase
 class Flavor(ModelBase):
     __tablename__: str = "flavors"
 
-    id: int = sa.Column(sa.BigInteger, primary_key=True, autoincrement=True)
+    id: int = sa.Column(sa.Integer, primary_key=True, autoincrement=True)
     created_at: datetime = sa.Column(sa.DateTime, default=datetime.now, index=True)
     name: str = sa.Column(sa.String(45), unique=True, nullable=False)
